cat-pi-monitor/requirements.txt
grpcio==1.34.1
picamera